# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Think Then Talk API
  description: The complete API specification for the Think Then Talk event platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: List Events
      description: Returns a list of events, with optional filters.
      parameters:
        - name: showPastEvents
          in: query
          schema:
            type: boolean
          description: Whether to include events that have already passed.
        - name: type
          in: query
          schema:
            type: string
            enum:
              - ONLINE
              - OFFLINE
          description: Filter by event type.
        - name: city
          in: query
          schema:
            type: string
          description: Filter by city for offline events.
        - name: category
          in: query
          schema:
            type: string
          description: Filter by a specific category.
        - name: "categories[]"
          in: "query"
          schema:
            type: "array"
            items:
              type: "string"
          style: "form"
          explode: true
          description: "Filter by multiple categories. Example: ?categories[]=design&categories[]=tech"
        - name: "dateRange[from]"
          in: "query"
          schema:
            type: "string"
            format: "date-time"
          description: "Filter events starting from this date/time (inclusive)."
        - name: "dateRange[to]"
          in: "query"
          schema:
            type: "string"
            format: "date-time"
          description: "Filter events up to this date/time (inclusive)."
      responses:
        '200':
          description: A list of events.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /events/{eventId}:
    get:
      tags:
        - Events
      summary: Get Event Details
      description: Returns the full details for a single event.
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
...
  '/admin/events/{eventId}':
    get:
      tags:
        - Admin
      summary: Get Event (Admin)
      description: Retrieves a single event for editing in the admin panel.
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
    put:
      tags:
        - Admin
      summary: Update Event (Admin)
      description: Updates the core details of an event.
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventFormData'
      responses:
        '200':
          description: Event updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found.
...
  /payments:
    post:
      tags:
        - Payments
      summary: Create Registration and Payment
      description: Creates a new event registration for the authenticated user and initiates a payment record. The frontend calls this single endpoint to start the checkout process.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentBody'
      responses:
        '201':
          description: Registration initiated, payment record created. Returns a payment object to redirect the user to the payment gateway.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Invalid input data (e.g., validation error on formData) or event capacity is full.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/payments/{paymentId}':
    get:
      tags:
        - Payments
      summary: Get Payment Details
      description: Retrieves the details and status of a single payment record. Used by the frontend on the callback page to check the status.
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Payment record not found.
    post:
      tags:
        - Payments
      summary: Verify Payment Status
      description: Endpoint for the frontend to confirm payment status after returning from the payment gateway.
      security:
        - bearerAuth: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum:
                    - SUCCESS
                    - FAILED
      responses:
        '200':
          description: Payment status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          description: Payment record not found.
...
  /admin/payments:
    get:
      tags:
        - Admin
      summary: List All Payments (Admin)
      description: Retrieves a list of all payment records for financial overview.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of payments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
...
  /admin/payments/export:
    get:
      tags:
        - Admin
      summary: 'Note: This endpoint is not currently called by the frontend; it is a planned/optional backend feature.'
      description: 'Note: This endpoint is not currently called by the frontend; it is a planned/optional backend feature.'
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum:
              - SUCCESS
              - PENDING
              - FAILED
      responses:
        '200':
          description: A CSV file of payments.
          content:
            text/csv:
              schema:
                type: string
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
      required:
        - message
...
    CreatePaymentBody:
      type: object
      description: The data required to initiate a payment for a registration.
      properties:
        eventId:
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        amount:
          type: number
        currency:
          type: string
          enum:
            - TOMAN
            - IRR
        formData:
          $ref: '#/components/schemas/RegistrationFormData'
      required:
        - eventId
        - ticketType
        - amount
        - currency
        - formData
    TicketType:
      type: string
      enum:
        - EARLY_BIRD
        - STANDARD
        - SUPPORTER
...
    RegistrationFormData:
      type: object
      description: The full set of data collected during user registration for an event.
      required:
        - firstNameFa
        - lastNameFa
        - mobile
        - isEmployed
        - acceptedRules
      properties:
        firstNameFa:
          type: string
        lastNameFa:
          type: string
        firstNameEn:
          type: string
        lastNameEn:
          type: string
        age:
          type: number
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
        mobile:
          type: string
          description: User's mobile number used during registration.
        email:
          type: string
          format: email
        educationLevel:
          type: string
        fieldOfStudy:
          type: string
        isEmployed:
          type: boolean
        jobTitle:
          type: string
        languageLevel:
          type: string
        referralSource:
          type: string
        referrerName:
          type: string
        otherReferralSource:
          type: string
        acceptedRules:
          type: boolean
...
    Event:
      type: object
      required:
        - id
        - title
        - summary
        - description
        - type
        - startDateTime
        - capacityTotal
        - capacityRemaining
        - showRemainingCapacity
        - categories
        - tickets
        - resources
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        description:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        type:
          type: string
          enum:
            - ONLINE
            - OFFLINE
        address:
          type: string
        city:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        capacityTotal:
          type: integer
        capacityRemaining:
          type: integer
        showRemainingCapacity:
          type: boolean
        categories:
          type: array
          items:
            type: string
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/EventTicketConfig'
        resources:
          type: array
          items:
            $ref: '#/components/schemas/EventResource'
        publicDiscountIds:
          type: array
          items:
            type: string
        posterUrl:
          type: string
          format: uri
    EventFormData:
      type: object
      required:
        - title
        - summary
        - description
        - type
        - startDateTime
        - capacityTotal
        - showRemainingCapacity
      properties:
        title:
          type: string
        categories:
          type: string
        summary:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        description:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        type:
          type: string
          enum:
            - ONLINE
            - OFFLINE
        city:
          type: string
        address:
          type: string
        startDateTime:
          type: string
          format: date-time
        capacityTotal:
          type: integer
        showRemainingCapacity:
          type: boolean
        publicDiscountIds:
          type: array
          items:
            type: string
        posterUrl:
          type: string
          format: uri
    EventResource:
      type: object
      required:
        - id
        - title
        - accessLevel
        - url
      properties:
        id:
          type: string
        title:
          type: object
          properties:
            fa:
              type: string
            en:
              type: string
        description:
          type: string
        accessLevel:
          type: string
          enum:
            - PUBLIC
            - REGISTERED_ONLY
        url:
          type: string
          format: uri
    EventTicketConfig:
      type: object
      required:
        - type
        - price
        - currency
        - quantityTotal
        - quantitySold
      properties:
        type:
          $ref: '#/components/schemas/TicketType'
        price:
          type: number
        currency:
          type: string
          enum:
            - IRR
            - TOMAN
        quantityTotal:
          type: integer
        quantitySold:
          type: integer
        earlyBirdEndDate:
          type: string
          format: date-time
    Payment:
      type: object
      required:
        - id
        - registrationId
        - eventId
        - ticketType
        - amount
        - currency
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        registrationId:
          type: string
        eventId:
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        amount:
          type: number
        currency:
          type: string
          enum:
            - TOMAN
            - IRR
        status:
          type: string
          enum:
            - PENDING
            - SUCCESS
            - FAILED
        gatewayTransactionId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT